Moved lib/money/money* to lib/
