+++ lib/money/formatting.rb
