+++ spec/money/bank/base_spec.rb
