+++ spec/money/bank/single_currency_spec.rb
