+++ spec/money/bank/variable_exchange_spec.rb
