+++ spec/money/currency/loader_spec.rb
