+++ spec/money/rates_store/memory_spec.rb
